+++ src/components/organisms/LinkCard/index.ts
